🔖 1.8.19: ♻ Refactor models + `LastCard` fields in `StudyMember`
